# Remove reflection from all namespaces (34% faster)