--- a/project.clj
+++ b/project.clj
@@ -10,7 +10,8 @@
                  [crypto-equality "1.0.0"]
                  [commons-codec/commons-codec "1.10"]]
   :profiles {:dev {:dependencies [[speclj "3.3.2"]
-                                  [org.bouncycastle/bcpkix-jdk15on "1.56"]]}}
+                                  [org.bouncycastle/bcpkix-jdk15on "1.56"]]
+                   :global-vars {*warn-on-reflection* true}}}
   :repositories [["releases" {:url "https://clojars.org/repo"
                               :sign-releases false }]]
   :plugins [[speclj "3.3.2"]]