--- a/src/jerks_whistling_tunes/utils.clj
+++ b/src/jerks_whistling_tunes/utils.clj
@@ -1,14 +1,14 @@
 (ns jerks-whistling-tunes.utils
   (:import org.apache.commons.codec.binary.Base64))
 
-(def ^:private base-64-encoder (Base64. -1 (.getBytes "") true))
+(def ^:private ^Base64 base-64-encoder (Base64. -1 (.getBytes "") true))
 
 (defn decode-base-64
   "Returns the decoded base 64 url safe encoded string."
-  [str]
-  (.decode base-64-encoder str))
+  [s]
+  (.decode base-64-encoder s))
 
 (defn encode-base-64
   "Returns a base 64 url safe encoded representation of the string"
-  [bytes]
-  (String. (.encode base-64-encoder bytes) "UTF-8"))
+  [^bytes bs]
+  (String. (.encode base-64-encoder bs) "UTF-8"))