Before

```text
Evaluation count : 7374 in 6 samples of 1229 calls.
             Execution time mean : 80.742711 µs
    Execution time std-deviation : 570.797426 ns
   Execution time lower quantile : 80.096170 µs ( 2.5%)
   Execution time upper quantile : 81.605858 µs (97.5%)
                   Overhead used : 6.696816 ns
```

After

```text
Evaluation count : 11454 in 6 samples of 1909 calls.
             Execution time mean : 53.017780 µs
    Execution time std-deviation : 665.405999 ns
   Execution time lower quantile : 52.061622 µs ( 2.5%)
   Execution time upper quantile : 53.609765 µs (97.5%)
                   Overhead used : 6.763715 ns
```

Using criterium and the following session (I've replaced the actual values with fillers, to hide necessary info -- the validation was successful though):

```clojure
user=> (use 'jerks-whistling-tunes.core)
nil
user=> (use 'jerks-whistling-tunes.sign)
nil
user=> (use 'jerks-whistling-tunes.utils)
nil
user=> (use 'criterium.core)
nil
user=> (quick-bench (validate "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX.YYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYY.ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ" (-> "CCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCC" hs256 signature) (aud "DDDD") exp))
```